split magic and netgen
