+++ siliconcompiler/tools/magic/drc.py
@@ -1,0 +1,18 @@
+
+from .magic import setup as setup_tool
+
+def setup(chip):
+    ''' Helper method to setup DRC-specific configs.
+    '''
+
+    # Generic tool setup
+    setup_tool(chip)
+
+    tool = 'magic'
+    step = chip.get('arg','step')
+    index = chip.get('arg','index')
+    task = chip.get_task(step, index)
+    design = chip.top()
+
+    report_path = f'reports/{design}.drc'
+    chip.set('tool', tool, 'task', task, 'report', step, index, 'drvs', report_path)
